Initial Commit with Articles
I imported blog articles in the posts folder with them being in markdown
format. The homepage(index.js) will have a features article section and a
recent articles section so people can see the latest articles first. At
the end will be a button to see all the articles, which will list all
articles in a different card format. The sites content is pretty much all
there, but need to make the site responsive.