Implemented the search input and added more posts.
Added another blog post to the website with image. I did some more styling
to make the website more responsive. Still need to make more improvements
towards the responsiveness of the website. Finally, implemented a search
api input that shows posts that match what is searched for on the articles
page.