Improvements to SEO, Search Function, and added Categories
There was a lot implemented in this update. The biggest feature was the
implementation of the Action Plan Categories that are accessible on the
articles page. This allows posts to have their own categories and can
filter posts using the category links. In addition, added new highlighting
when typing in the search input to give a better search experience for the
user. Lastly, added a lot of SEO improvements like robot.txt, meta tags,
favicons, and page titles. Still need to clean up code in a refactoring
session and add a categories description.