Moved Category Sidebar and Implemented a Category Tag on Posts
Implemented the CategoryLabel.js that takes the categories and displays
the proper category on each post. You can even click it to take you to all
articles associated with the category. Furthermore, I remade the categories
sidebar and put it at the top middle of the articles page. Lastly, added
new favicon and did some responsive css additions to the website. I would
like to add a tooltip to the categories in future updates.